pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tasks-backend"
version = "0.1.0"
dependencies = [
    "bcrypt==4.2.1",
    "fastapi[standard]>=0.115.6",
    "passlib==1.7.4",
    "psycopg2>=2.9.10",
    "pyjwt==2.10.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22"
]
description = "Backend server for tasks app using FastAPI and SQLModel"
readme = "README.md"
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.2"
]

[project.scripts]
run = "tasks_backend.run:run_server"
